Baked in mechanism for Butler.get from S3 storage for simplest of Datasets with no Dimensions.
Changes
-------
Added path parsing utilities that figure out if a path is an S3 URI or filesystem path
to the utilities.

ButlerConfig checks at instantiation, in the case its not being instantiated from another
ButlerConfig instance, whether the given path is an S3 URI. If it is, the butler yaml
configuration file is downloaded and a ButlerConfig is instantiated from that. The
download location is hard-coded and won't work on different machine.

Edited fromConfig instantiation method of Registry. If a registry is instantiated from
general Config class it checks whether the path to db is an S3 URI or filesystem path.
If it's S3 URI it will download the database locally to hardcoded path that will only
work on my machine. The remainder of instantiation is then executed as normal from the
downloaded file. The RegistryConfig is updated to reflect the local downloaded path.

S3Datastore has a different config than the default PosixDatastore one. It manually
appends the s3:// to default datastore templates. I do not know outfront if this is
really required or not.

Initialization of the S3DataStore is different compared to PosixDatastore. Attributes
service, bucket and rootpath are added identifying the 's3://', bucket name and path
to root of the repository elements of the object storage repository respectively.
Attributes session and client provide access, through boto3 API, to upload and download
functionality.
The file exists and filesize checks were changed to query the S3 API. Currently this is
performed by making a listing request to the S3 Bucket. This is allegedly the fastest,
but for searching for the name of the file. It is also 12.5 times as expensive as directly
making a request for the object and catching the exception. I do not know if there are
data that can have multiple different paths returned. I suspect not, so this might not
be worth it.
There is some functionality that will create a bucket if one doesn't exists in the
S3DataStore init method, but this is just showcase code - doesn't work or make sense
because Registry has to be instantiated before Datastore is and we need that yaml config
file to exists - so the bucket must exist too.
The get method was not particularily drastically changed, because the  mechanism was pushed
to formatters.

LocationFactory used in the DataStores was changed to return S3Location if it was
instantiated from an S3 URI.

The S3Location keeps the repository root, the bucket name and the full original bucket
URI stored as attributes so its possible to figure out what is the bucket path from
which to download the file in the Formatters.

The S3FitsExposureFormatter readFull method uses the S3Location to download the exposure
to a baked in path that will, again, only work for me. The mapped DatasetType is then used
to instantiate from that file and return appropriate in-memory (hopefully) object. All
downloaded Datasets are given the same name to avoid clutter, but this should be done
differently anyhow so the hacking solution will do for now.